+++ unix_network_programming/server/pthread09.c
@@ -1,8 +1,9 @@
 #include	"unpthread.h"
 #include	"pthread09.h"
+#include <inttypes.h>
 
 void
-thread_make(int i)
+thread_make(intptr_t i)
 {
 	void	*thread_main(void *);
 
@@ -20,11 +21,11 @@ thread_main(void *arg)
 
 	cliaddr = Malloc(addrlen);
 
-	printf("thread %d starting\n", (int) arg);
+	printf("thread %"PRIdPTR" starting\n", (intptr_t) arg);
 	for ( ; ; ) {
 		clilen = addrlen;
 		connfd = Accept(listenfd, cliaddr, &clilen);
-		tptr[(int) arg].thread_count++;
+		tptr[(intptr_t) arg].thread_count++;
 
 		web_child(connfd);		/* process the request */
 		Close(connfd);
