+++ unix_network_programming/server/serv07.c
@@ -7,8 +7,8 @@ pthread_mutex_t	mlock = PTHREAD_MUTEX_INITIALIZER;
 int
 main(int argc, char **argv)
 {
-	int		i;
-	void	sig_int(int), thread_make(int);
+	intptr_t i;
+	void	sig_int(int), thread_make(intptr_t);
 
 	if (argc == 3)
 		listenfd = Tcp_listen(NULL, argv[1], &addrlen);
